app/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
click==6.7
flask==0.12.2
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
psycopg2-binary==2.7.4
werkzeug==0.14.1
